requirements.txt
streamlit==1.31.1
jira==3.5.2
pandas==2.2.0
python-dotenv==1.0.1
plotly==5.18.0 